src/server/alo-cidadao/requirements.txt
Flask==2.2.2
Flask-PyMongo==2.3.0
pymongo==4.2.0
SQLAlchemy==1.4.40
dnspython==2.2.1
Flask-Cors==3.0.10
Pillow==9.2.0